requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Utilities
django-filter==23.3
Pillow==10.1.0
django-extensions==3.2.3

# Development
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Data validation
email-validator==2.1.0

# API Documentation
drf-yasg==1.21.7




